Add files via upload
The Simple Calculator is a basic web application that allows users to perform simple arithmetic calculations. It is implemented using HTML for the structure, CSS for the styling, and JavaScript for the functionality. The calculator features a user-friendly interface with number buttons, operator buttons (+, -, *, /), and an equal (=) button to execute calculations. When users click on the buttons, the corresponding numbers and operators are displayed on the calculator screen. Upon clicking the equal button, the calculator evaluates the expression and displays the result. This straightforward implementation provides a convenient tool for performing basic mathematical calculations right in the browser.